+++ src/main/java/com/newworld/saegil/ai/config/TtsProvider.java
@@ -1,4 +1,4 @@
-package com.newworld.saegil.llm.config;
+package com.newworld.saegil.ai.config;
 
 public enum TtsProvider {
     OPENAI, ELEVENLABS
