+++ src/main/java/com/newworld/saegil/ai/controller/TextToSpeechRequest.java
@@ -1,0 +1,6 @@
+package com.newworld.saegil.ai.controller;
+
+import com.newworld.saegil.ai.config.TtsProvider;
+
+public record TextToSpeechRequest(String text, TtsProvider provider) {
+}
