+++ src/main/java/com/newworld/saegil/ai/service/OpenAiChat.java
@@ -1,4 +1,4 @@
-package com.newworld.saegil.llm.service;
+package com.newworld.saegil.ai.service;
 
 import org.springframework.web.multipart.MultipartFile;
