+++ src/main/java/com/newworld/saegil/ai/service/TextToSpeechService.java
@@ -1,7 +1,7 @@
-package com.newworld.saegil.llm.service;
+package com.newworld.saegil.ai.service;
 
-import com.newworld.saegil.llm.config.ProxyProperties;
-import com.newworld.saegil.llm.config.TtsProvider;
+import com.newworld.saegil.ai.config.ProxyProperties;
+import com.newworld.saegil.ai.config.TtsProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.openai.OpenAiAudioSpeechModel;
