--- src/main/java/com/newworld/saegil/llm/controller/TextToSpeechRequest.java
@@ -1,6 +1,0 @@
-package com.newworld.saegil.llm.controller;
-
-import com.newworld.saegil.llm.config.TtsProvider;
-
-public record TextToSpeechRequest(String text, TtsProvider provider) {
-}
